Rename bids_conversion module to match command name.
